fixes from 42workstation

+++ push_swap/Includes/Libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: yhruda <yhruda@student.42warsaw.pl>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 19:40:54 by admin             #+#    #+#             */
-/*   Updated: 2025/09/05 21:34:44 by yhruda           ###   ########.fr       */
+/*   Updated: 2025/09/06 12:00:53 by yhruda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	i;
